Refactor unthorized access
